Fix logger
