+++ src/main.py
@@ -1,12 +1,17 @@
 import os
 import signal
 import logging
-import multiprocessing
 import uvicorn
+import multiprocessing
+
+from fastapi import Request
+from fastapi.responses import HTMLResponse, RedirectResponse
+from fastapi.staticfiles import StaticFiles
+from fastapi.templating import Jinja2Templates
 
 from module import app
 from module.api import router
-from module.conf import VERSION, setup_logger, settings
+from module.conf import VERSION, settings
 
 
 logger = logging.getLogger(__name__)
@@ -14,7 +19,7 @@ logger = logging.getLogger(__name__)
 main_process = multiprocessing.Process(target=app.run)
 
 
-@router.get("/api/v1/restart")
+@router.get("/api/v1/restart", tags=["program"])
 async def restart():
     global main_process
     logger.info("Restarting...")
@@ -25,18 +30,24 @@ async def restart():
     return {"status": "success"}
 
 
-def show_info():
-    with open("icon", "r") as f:
-        for line in f.readlines():
-            logger.info(line.strip("\n"))
-    logger.info(f"Version {VERSION}  Author: EstrellaXD Twitter: https://twitter.com/Estrella_Pan")
-    logger.info("GitHub: https://github.com/EstrellaXD/Auto_Bangumi/")
-    logger.info("Starting AutoBangumi...")
+if VERSION != "DEV_VERSION":
+    router.mount("/assets", StaticFiles(directory="templates/assets"), name="assets")
+    templates = Jinja2Templates(directory="templates")
+
+    # HTML Response
+    @router.get("/{full_path:path}", response_class=HTMLResponse, tags=["html"])
+    def index(request: Request):
+        context = {"request": request}
+        return templates.TemplateResponse("index.html", context)
+else:
+    @router.get("/", status_code=302, tags=["html"])
+    def index():
+        return RedirectResponse("/docs")
 
 
 if __name__ == "__main__":
-    setup_logger()
-    show_info()
+    log_config = uvicorn.config.LOGGING_CONFIG
+    log_config["formatters"]["default"]["fmt"] = "[%(asctime)s] %(levelname)-8s  %(message)s"
     main_process.start()
-    uvicorn.run(router, host="0.0.0.0", port=settings.program.webui_port)
+    uvicorn.run(router, host="0.0.0.0", port=settings.program.webui_port, log_config=log_config)
