+++ src/module/api.py
@@ -1,30 +1,29 @@
-import uvicorn
-from uvicorn.config import LOGGING_CONFIG
-from fastapi import FastAPI, Request
-from fastapi.responses import HTMLResponse, FileResponse, RedirectResponse
-from fastapi.templating import Jinja2Templates
-from fastapi.staticfiles import StaticFiles
 import logging
+from fastapi import FastAPI
+from fastapi.responses import FileResponse
 
 from module.core import APIProcess
-from module.conf import settings, DATA_PATH, LOG_PATH, VERSION
+from module.conf import DATA_PATH, LOG_PATH, settings
 from module.utils import json_config
 from module.models.api import *
 from module.models import Config
 
-logger = logging.getLogger(__name__)
 
 router = FastAPI()
 api_func = APIProcess()
 
-if VERSION != "DEV_VERSION":
-    router.mount("/assets", StaticFiles(directory="templates/assets"), name="assets")
-    templates = Jinja2Templates(directory="templates")
 
+@router.on_event("startup")
+async def startup_event():
+    logger = logging.getLogger("uvicorn.access")
+    handler = logging.StreamHandler()
+    handler.setLevel(logging.INFO)
+    handler.setFormatter(logging.Formatter("[%(asctime)s] %(levelname)-8s  %(message)s"))
+    logger.addHandler(handler)
 
 
 @router.get("/api/v1/data")
-def get_data():
+async def get_data():
     data = json_config.load(DATA_PATH)
     return data
 
@@ -62,22 +61,4 @@ async def add_rule(info: AddRule):
 @router.post("/api/v1/updateConfig", tags=["config"])
 async def update_config(config: Config):
     return api_func.update_config(config)
-
-if VERSION != "DEV_VERSION":
-# HTML Response
-    @router.get("/{full_path:path}", response_class=HTMLResponse)
-    def index(request: Request):
-        context = {"request": request}
-        return templates.TemplateResponse("index.html", context)
-else:
-    @router.get("/", status_code=302)
-    def index():
-        return RedirectResponse("/docs")
-
-
-def run():
-    LOGGING_CONFIG["formatters"]["default"]["fmt"] = "[%(asctime)s] %(levelprefix)s %(message)s"
-    uvicorn.run(router, host="0.0.0.0", port=settings.program.webui_port)
-
-
